Error caused when querying for an A record when CNAME exists
foo.bar.net: CNAME->baz.net

If you query for an A record for the above domain, it's going to end up looking also for
a CNAME record (since no A exists), however if you send +norecurse it's not going to match
the if(recurse && cname) clause. This caused it to match the other clause, which means we'd be
assigning a host string "baz.net" to an A resource record. This caused an error that wasn't
being checked for.